+++ adventure/Creature.py
@@ -1,42 +1,28 @@
+from math import nan
 
-class Creature:
-    name = None                         # Attributes!
-    description = None
-    current_location = None
-    items = []
 
-    def __init__(self, attributes):     # Methods! 
+class Creature:                                # this is the main class, that the other creatures derive stuff from
+    name: None
+    description: None
+    health: nan
+    location: None
+    inventory: []
 
-        def set_attribute_for_self(key, value):
-            setattr(self, key, value)
-
-        map(set_attribute_for_self, attributes.keys(), attributes.values())
-
-        #---------
-
-        for key, value in attributes.items():
-            setattr(self, key, value)
-
-    def describe(self):
-        print(self.description)
+    hostile: None
 
 
 class Dragon(Creature):
-    is_sleep = True
-
-    def speak(self):
-        print("Roar! My name is "+ self.name)
+    sleeping = True
 
 
-class Player(Creature):
-    def move(self, new_location):
-        # check if the player can move to new location
-        self.current_location = new_location
+class Player(Creature):                             # Note: jos class on tyhjä, kirjoita sinne 'pass', muuten tulee virhe :)
+    pass
 
-#------------------------
-# In the Main file:
+class Rat(Creature):
+    hostile = False
 
-player = Player({ "name": "Joakim", "location": "red" })
-print(player.current_location)
-player.move("green")
-print(player.current_location)
+class BossRat(Rat):
+    hostile = True
+
+
+
